python-twootfeed/requirements.txt
apipkg==1.4
astroid==1.6.2
attrs==17.4.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
execnet==1.5.0
feedgenerator==1.9
flake8==3.5.0
Flask==0.12.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
Mastodon.py==1.1.2
mccabe==0.6.1
more-itertools==4.1.0
oauthlib==2.0.7
pluggy==0.6.0
py==1.5.3
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.3
pytest==3.5.0
pytest-cache==1.0
pytest-flake8==1.0.0
pytest-isort==0.1.0
python-dateutil==2.7.2
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
